requirements.txt
flask==2.1.0
flask-cors==3.0.10
hypercorn==0.14.3
python-dotenv==0.19.2
pyyaml==6.0
tinydb==4.7.0
prompt-toolkit==3.0.28
groq==0.3.0
aiohttp==3.8.3
cerebras-cloud-sdk==0.2.0